// britv3.0/src/types/marketplace.ts
export type ServiceCategory =
  | "conveyancing"
  | "surveying"
  | "mortgage_broker"
  | "moving_company"
  | "home_inspector"
  | "cleaning"
  | "handyman"
  | "plumber"
  | "electrician"
  | "landscaping"
  | "interior_design"
  | "architect"
  | "property_management"
  | "pest_control"
  | "locksmith"
  | "other";